Some changes to the Makefile and FreeRTOSConfig.h
The big thing is some additions to these files.  The SET_BIT (macro?  this is a #define made global function... )and ZERO_BIT have been
re-added.  Bool has been put back in.  These are in the FreeRTOSConfig.h.  The Makefile has been changed to include the GPIO stuff.
The GPIO stuff apparently is untracked right now.  Not sure why...