Update for ADC functionality.  ADC should work but is untested.
ADC should work at this point.  Some register checking and testing remains, but the code compiles and has been largely checked for correct
logic.  A primary assumption is that the ADC logic is largely the same for ADC handling (register bit shifts, etc) from the LPC2148 to LPC23xx
and this has held true in all cases seen so far.  Makefile has been modified to account for specifying LPC2368 or LPC2378, and GPIO and
main have been modified to account for a failure of the previous processor selection method.