Added functionality to handle repeated starts, arbitrary stops and stop starts in I2C communication
Modifications rely on and monitor flags in next significant byte in "TransmitData" such that flags (start and
stop) are mirrored in this second least significant byte, and no data is in the least significant byte.  Thus
the array TransmitData should contain only values (control characters?) 0x3000, 0x2000, 0x1000 and byte data
(0xFF - 0x00).

0x3000	Send a stop then start
0x2000	Send a stop
0x1000	Send a repeated start
The following scheme may be used but is not currently implemented:

0x3100	Send a stop then start, BUT ACK was expected!  if not recieved, do something
0x3200	Send a stop then start, BUT ACK was NOT expected!  if recieved, do something
0x2200
0x2100
0x1200
0x1100

the dataLength array should treat a stop, a repeated start or a stop/start as a data item with a length of 1.
For each such element additional to the data bytes, the dataLength should be incremented by 1, so that the
state transitions on these control actions may be counted.

The portions of the ISR that handle data transmissions should also handle control characters.  It is hoped
that these are only states 0x18, 0x20, 0x28, 0x30 and 0x38.  Only 0x18 and 0x28 (dealing with transmissions
that have been acknowledged by the slave) have been modified at this time.