added directory USBAllTest1 It contains host and device code. the host directory also contains a copy of libusb-1.0.8 Both the host and device code is a derivative of LPCUSBIntegrationTesting and USBTestHostLibs respectively. Now it is under one roof.
The device code was modified by commenting out sections of code that blinked the STAT LED.
It seemed there were multiple sections of code competing for the honor, making it hard to tell why the LED was blinking.
so now only the packets recieved from the host that explicitly set the LED actually do so.

The host code is a work in progress. I am cleaning up the code, and experimenting with a more sophisticated event pump
that resides in its own thread(libusb_handle_events).

On my machine, the LED on the dev board blinks really fast, but is totally visible.
this may not be the case on a different setup.